Make Impersonator an interface
