+++ cdap-app-fabric/src/test/java/co/cask/cdap/app/runtime/AbstractProgramRuntimeServiceTest.java
@@ -25,7 +25,7 @@ import co.cask.cdap.common.ArtifactNotFoundException;
 import co.cask.cdap.common.app.RunIds;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.io.Locations;
-import co.cask.cdap.common.security.Impersonator;
+import co.cask.cdap.common.security.DefaultImpersonator;
 import co.cask.cdap.common.utils.Tasks;
 import co.cask.cdap.internal.app.runtime.ProgramControllerServiceAdapter;
 import co.cask.cdap.internal.app.runtime.SimpleProgramOptions;
@@ -74,7 +74,7 @@ public class AbstractProgramRuntimeServiceTest {
 
     final ProgramRuntimeService runtimeService =
       new AbstractProgramRuntimeService(CConfiguration.create(), runnerFactory, null,
-                                        new Impersonator(CConfiguration.create(), null, null)) {
+                                        new DefaultImpersonator(CConfiguration.create(), null, null)) {
       @Override
       public ProgramLiveInfo getLiveInfo(Id.Program programId) {
         return new ProgramLiveInfo(programId, "runtime") { };
@@ -282,7 +282,8 @@ public class AbstractProgramRuntimeServiceTest {
 
     protected TestProgramRuntimeService(CConfiguration cConf, ProgramRunnerFactory programRunnerFactory,
                                         @Nullable ArtifactRepository artifactRepository, RuntimeInfo extraInfo) {
-      super(cConf, programRunnerFactory, artifactRepository, new Impersonator(CConfiguration.create(), null, null));
+      super(cConf, programRunnerFactory, artifactRepository,
+            new DefaultImpersonator(CConfiguration.create(), null, null));
       this.extraInfo = extraInfo;
     }
