+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/app/deploy/ConfiguratorTest.java
@@ -25,7 +25,7 @@ import co.cask.cdap.app.runtime.DummyProgramRunnerFactory;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.common.guice.ConfigModule;
-import co.cask.cdap.common.security.Impersonator;
+import co.cask.cdap.common.security.DefaultImpersonator;
 import co.cask.cdap.internal.app.ApplicationSpecificationAdapter;
 import co.cask.cdap.internal.app.deploy.pipeline.NamespacedImpersonator;
 import co.cask.cdap.internal.app.runtime.artifact.ArtifactRepository;
@@ -91,17 +91,17 @@ public class ConfiguratorTest {
     LocationFactory locationFactory = new LocalLocationFactory(TMP_FOLDER.newFolder());
     Location appJar = AppJarHelper.createDeploymentJar(locationFactory, WordCountApp.class);
     Id.Artifact artifactId = Id.Artifact.from(Id.Namespace.DEFAULT, WordCountApp.class.getSimpleName(), "1.0.0");
+    CConfiguration cConf = CConfiguration.create();
     ArtifactRepository artifactRepo = new ArtifactRepository(conf, null, null, authorizer,
                                                              new DummyProgramRunnerFactory(),
-                                                             new Impersonator(CConfiguration.create(), null, null),
+                                                             new DefaultImpersonator(cConf, null, null),
                                                              authEnforcer, authenticationContext);
 
     // Create a configurator that is testable. Provide it a application.
     try (CloseableClassLoader artifactClassLoader =
-           artifactRepo.createArtifactClassLoader(appJar,
-                                                  new NamespacedImpersonator(artifactId.getNamespace().toEntityId(),
-                                                                             new Impersonator(CConfiguration.create(),
-                                                                                              null, null)))) {
+           artifactRepo.createArtifactClassLoader(
+             appJar, new NamespacedImpersonator(artifactId.getNamespace().toEntityId(),
+                                                new DefaultImpersonator(cConf, null, null)))) {
       Configurator configurator = new InMemoryConfigurator(conf, Id.Namespace.DEFAULT, artifactId,
                                                            WordCountApp.class.getName(), artifactRepo,
                                                            artifactClassLoader, null, "");
@@ -124,18 +124,18 @@ public class ConfiguratorTest {
     LocationFactory locationFactory = new LocalLocationFactory(TMP_FOLDER.newFolder());
     Location appJar = AppJarHelper.createDeploymentJar(locationFactory, ConfigTestApp.class);
     Id.Artifact artifactId = Id.Artifact.from(Id.Namespace.DEFAULT, ConfigTestApp.class.getSimpleName(), "1.0.0");
+    CConfiguration cConf = CConfiguration.create();
     ArtifactRepository artifactRepo = new ArtifactRepository(conf, null, null, authorizer,
                                                              new DummyProgramRunnerFactory(),
-                                                             new Impersonator(CConfiguration.create(), null, null),
+                                                             new DefaultImpersonator(cConf, null, null),
                                                              authEnforcer, authenticationContext);
 
     ConfigTestApp.ConfigClass config = new ConfigTestApp.ConfigClass("myStream", "myTable");
     // Create a configurator that is testable. Provide it a application.
     try (CloseableClassLoader artifactClassLoader =
-           artifactRepo.createArtifactClassLoader(appJar,
-                                                  new NamespacedImpersonator(artifactId.getNamespace().toEntityId(),
-                                                                             new Impersonator(CConfiguration.create(),
-                                                                                              null, null)))) {
+           artifactRepo.createArtifactClassLoader(
+             appJar, new NamespacedImpersonator(artifactId.getNamespace().toEntityId(),
+                                                new DefaultImpersonator(cConf, null, null)))) {
       Configurator configuratorWithConfig =
         new InMemoryConfigurator(conf, Id.Namespace.DEFAULT, artifactId, ConfigTestApp.class.getName(),
                                  artifactRepo, artifactClassLoader, null, new Gson().toJson(config));
