+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/app/runtime/artifact/ArtifactInspectorTest.java
@@ -24,7 +24,7 @@ import co.cask.cdap.common.InvalidArtifactException;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.common.io.Locations;
-import co.cask.cdap.common.security.Impersonator;
+import co.cask.cdap.common.security.DefaultImpersonator;
 import co.cask.cdap.common.utils.DirUtils;
 import co.cask.cdap.internal.app.deploy.pipeline.NamespacedImpersonator;
 import co.cask.cdap.internal.app.runtime.artifact.app.InvalidConfigApp;
@@ -77,10 +77,10 @@ public class ArtifactInspectorTest {
     Id.Artifact artifactId = Id.Artifact.from(Id.Namespace.DEFAULT, "InvalidConfigApp", "1.0.0");
     Location artifactLocation = Locations.toLocation(appFile);
     try (CloseableClassLoader artifactClassLoader =
-           classLoaderFactory.createClassLoader(artifactLocation,
-                                                new NamespacedImpersonator(artifactId.getNamespace().toEntityId(),
-                                                                           new Impersonator(CConfiguration.create(),
-                                                                                            null, null)))) {
+           classLoaderFactory.createClassLoader(
+             artifactLocation, new NamespacedImpersonator(artifactId.getNamespace().toEntityId(),
+                                                          new DefaultImpersonator(CConfiguration.create(),
+                                                                                  null, null)))) {
       artifactInspector.inspectArtifact(artifactId, appFile, artifactClassLoader);
     }
   }
@@ -95,10 +95,10 @@ public class ArtifactInspectorTest {
     Id.Artifact artifactId = Id.Artifact.from(Id.Namespace.DEFAULT, "InspectionApp", "1.0.0");
     Location artifactLocation = Locations.toLocation(appFile);
     try (CloseableClassLoader artifactClassLoader =
-           classLoaderFactory.createClassLoader(artifactLocation,
-                                                new NamespacedImpersonator(artifactId.getNamespace().toEntityId(),
-                                                                           new Impersonator(CConfiguration.create(),
-                                                                                            null, null)))) {
+           classLoaderFactory.createClassLoader(
+             artifactLocation, new NamespacedImpersonator(artifactId.getNamespace().toEntityId(),
+                                                          new DefaultImpersonator(CConfiguration.create(),
+                                                                                  null, null)))) {
 
       ArtifactClasses classes = artifactInspector.inspectArtifact(artifactId, appFile, artifactClassLoader);
