+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/app/runtime/artifact/ArtifactStoreTest.java
@@ -25,7 +25,7 @@ import co.cask.cdap.api.plugin.PluginPropertyField;
 import co.cask.cdap.common.ArtifactAlreadyExistsException;
 import co.cask.cdap.common.ArtifactNotFoundException;
 import co.cask.cdap.common.conf.CConfiguration;
-import co.cask.cdap.common.security.Impersonator;
+import co.cask.cdap.common.security.DefaultImpersonator;
 import co.cask.cdap.internal.AppFabricTestHelper;
 import co.cask.cdap.internal.app.deploy.pipeline.NamespacedImpersonator;
 import co.cask.cdap.internal.app.runtime.artifact.app.inspection.InspectionApp;
@@ -1048,6 +1048,6 @@ public class ArtifactStoreTest {
     String contents) throws ArtifactAlreadyExistsException, IOException, WriteConflictException {
     artifactStore.write(artifactId, meta, ByteStreams.newInputStreamSupplier(Bytes.toBytes(contents)),
                         new NamespacedImpersonator(artifactId.getNamespace().toEntityId(),
-                                                   new Impersonator(CConfiguration.create(), null, null)));
+                                                   new DefaultImpersonator(CConfiguration.create(), null, null)));
   }
 }
