+++ cdap-data-fabric-tests/src/test/java/co/cask/cdap/data/stream/StreamDataFileTestBase.java
@@ -18,6 +18,7 @@ package co.cask.cdap.data.stream;
 import co.cask.cdap.api.flow.flowlet.StreamEvent;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.io.Locations;
+import co.cask.cdap.common.security.DefaultImpersonator;
 import co.cask.cdap.common.security.Impersonator;
 import co.cask.cdap.common.security.UnsupportedUGIProvider;
 import co.cask.cdap.data.file.FileReader;
@@ -78,7 +79,7 @@ public abstract class StreamDataFileTestBase {
 
   protected static CConfiguration cConf = CConfiguration.create();
 
-  private static final Impersonator impersonator = new Impersonator(cConf, new UnsupportedUGIProvider(), null);
+  private static final Impersonator impersonator = new DefaultImpersonator(cConf, new UnsupportedUGIProvider(), null);
 
   @Test
   public void testEmptyFile() throws Exception {
