+++ cdap-data-fabric-tests/src/test/java/co/cask/cdap/data/stream/TimePartitionedStreamTestBase.java
@@ -18,6 +18,7 @@ package co.cask.cdap.data.stream;
 import co.cask.cdap.api.flow.flowlet.StreamEvent;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.io.Locations;
+import co.cask.cdap.common.security.DefaultImpersonator;
 import co.cask.cdap.common.security.Impersonator;
 import co.cask.cdap.common.security.UnsupportedUGIProvider;
 import co.cask.cdap.proto.id.NamespaceId;
@@ -46,7 +47,7 @@ public abstract class TimePartitionedStreamTestBase {
 
   protected abstract LocationFactory getLocationFactory();
 
-  private static final Impersonator impersonator = new Impersonator(cConf, new UnsupportedUGIProvider(), null);
+  private static final Impersonator impersonator = new DefaultImpersonator(cConf, new UnsupportedUGIProvider(), null);
 
   @Test
   public void testTimePartition() throws IOException {
