+++ cdap-watchdog/src/test/java/co/cask/cdap/logging/write/LogCleanupTest.java
@@ -27,6 +27,7 @@ import co.cask.cdap.common.namespace.NamespaceQueryAdmin;
 import co.cask.cdap.common.namespace.NamespacedLocationFactory;
 import co.cask.cdap.common.namespace.NamespacedLocationFactoryTestClient;
 import co.cask.cdap.common.namespace.SimpleNamespaceQueryAdmin;
+import co.cask.cdap.common.security.DefaultImpersonator;
 import co.cask.cdap.common.security.Impersonator;
 import co.cask.cdap.common.security.UGIProvider;
 import co.cask.cdap.common.security.UnsupportedUGIProvider;
@@ -134,7 +135,7 @@ public class LogCleanupTest {
     txManager = injector.getInstance(TransactionManager.class);
     txManager.startAndWait();
     rootLocationFactory = injector.getInstance(RootLocationFactory.class);
-    impersonator = injector.getInstance(Impersonator.class);
+    impersonator = injector.getInstance(DefaultImpersonator.class);
   }
 
   @AfterClass
